Split up the mapping logic into multiple and more digestable files
