+++ Sources/TuistLoader/Models+ManifestMappers/Configuration+ManifestMapper.swift
@@ -1,0 +1,16 @@
+import Basic
+import Foundation
+import ProjectDescription
+import TuistCore
+import TuistSupport
+
+extension TuistCore.Configuration {
+    static func from(manifest: ProjectDescription.Configuration?, generatorPaths: GeneratorPaths) throws -> TuistCore.Configuration? {
+        guard let manifest = manifest else {
+            return nil
+        }
+        let settings = manifest.settings.mapValues(TuistCore.SettingValue.from)
+        let xcconfig = try manifest.xcconfig.flatMap { try generatorPaths.resolve(path: $0) }
+        return Configuration(settings: settings, xcconfig: xcconfig)
+    }
+}
