+++ Sources/TuistLoader/Models+ManifestMappers/Package+ManifestMapper.swift
@@ -1,0 +1,37 @@
+import Basic
+import Foundation
+import ProjectDescription
+import TuistCore
+import TuistSupport
+
+extension TuistCore.Package {
+    static func from(manifest: ProjectDescription.Package,
+                     path _: AbsolutePath,
+                     generatorPaths: GeneratorPaths) throws -> TuistCore.Package {
+        switch manifest {
+        case let .local(path: local):
+            return .local(path: try generatorPaths.resolve(path: local))
+        case let .remote(url: url, requirement: version):
+            return .remote(url: url, requirement: .from(manifest: version))
+        }
+    }
+}
+
+extension TuistCore.Package.Requirement {
+    static func from(manifest: ProjectDescription.Package.Requirement) -> TuistCore.Package.Requirement {
+        switch manifest {
+        case let .branch(branch):
+            return .branch(branch)
+        case let .exact(version):
+            return .exact(version.description)
+        case let .range(from, to):
+            return .range(from: from.description, to: to.description)
+        case let .revision(revision):
+            return .revision(revision)
+        case let .upToNextMajor(version):
+            return .upToNextMajor(version.description)
+        case let .upToNextMinor(version):
+            return .upToNextMinor(version.description)
+        }
+    }
+}
