+++ Sources/TuistLoader/Models+ManifestMappers/RunAction+ManifestMapper.swift
@@ -1,0 +1,22 @@
+import Basic
+import Foundation
+import ProjectDescription
+import TuistCore
+
+extension TuistCore.RunAction {
+    static func from(manifest: ProjectDescription.RunAction,
+                     generatorPaths: GeneratorPaths) throws -> TuistCore.RunAction {
+        let configurationName = manifest.configurationName
+        let arguments = manifest.arguments.map { TuistCore.Arguments.from(manifest: $0) }
+
+        var executableResolved: TuistCore.TargetReference?
+        if let executable = manifest.executable {
+            executableResolved = TargetReference(projectPath: try generatorPaths.resolveSchemeActionProjectPath(executable.projectPath),
+                                                 name: executable.targetName)
+        }
+
+        return RunAction(configurationName: configurationName,
+                         executable: executableResolved,
+                         arguments: arguments)
+    }
+}
