+++ Sources/TuistLoader/Models+ManifestMappers/SettingValue+ManifestMapper.swift
@@ -1,0 +1,16 @@
+import Basic
+import Foundation
+import ProjectDescription
+import TuistCore
+import TuistSupport
+
+extension TuistCore.SettingValue {
+    static func from(manifest: ProjectDescription.SettingValue) -> TuistCore.SettingValue {
+        switch manifest {
+        case let .string(value):
+            return .string(value)
+        case let .array(value):
+            return .array(value)
+        }
+    }
+}
